+++ Pandora/Paradigm/Structure/Ability/Morphable.hs
@@ -17,3 +17,6 @@ morph = morphing . TU . Tag @f
 data Walk a = Preorder a | Inorder a | Postorder a | Levelorder a
 
 data Morph a = Rotate a
+
+rotate :: forall f t . Morphable (Rotate f) t => t ~> Morphing (Rotate f) t
+rotate = morphing . TU . Tag @(Rotate f)