+++ CHANGELOG.md
@@ -350,3 +350,4 @@
 * Create `Some` umbrella module for certain data structures
 * Define experimental `~~>` type synonyms
 * Rename `/=` method of `Setoid` typeclass to `!=`
+* Define `rotate` method of `Morphable` typeclass